requirements-docker.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# AI/ML Models
dashscope>=1.20.0

# Data Validation & Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
PyJWT>=2.8.0

# Cloud Services
alibabacloud-videoenhan20200320>=1.0.0
alibabacloud-tea-openapi>=0.3.0
alibabacloud-tea-util>=0.3.0
oss2>=2.18.0
